requirements.txt
sqlmodel
apache-airflow
sqlalchemy
pypdf
confluent-kafka
tika
mariadb
sqlescapy
numpy
sentence_transformers
mysql.connector
psycopg2-binary
torch
spacy
ruamel.yaml